+++ src/commands/readObject.js
@@ -5,7 +5,7 @@ import { GitAnnotatedTag } from '../models/GitAnnotatedTag.js'
 import { GitCommit } from '../models/GitCommit.js'
 import { E, GitError } from '../models/GitError.js'
 import { GitTree } from '../models/GitTree.js'
-import { readObject as _readObject } from '../utils/readObject.js'
+import { readObject as _readObject } from '../storage/readObject.js'
 import { resolveTree } from '../utils/resolveTree.js'
 import { cores } from '../utils/plugins.js'
