+++ src/commands/sign.js
@@ -4,8 +4,8 @@ import { GitRefManager } from '../managers/GitRefManager.js'
 import { FileSystem } from '../models/FileSystem.js'
 import { E, GitError } from '../models/GitError.js'
 import { SignedGitCommit } from '../models/SignedGitCommit.js'
-import { readObject } from '../utils/readObject.js'
-import { writeObject } from '../utils/writeObject.js'
+import { readObject } from '../storage/readObject.js'
+import { writeObject } from '../storage/writeObject.js'
 import { cores } from '../utils/plugins.js'
 
 /**
