+++ src/internal-apis.js
@@ -26,6 +26,9 @@ export * from './models/GitSideBand'
 export * from './models/GitTree'
 export * from './models/SignedGitCommit'
 
+export * from './storage/readObject'
+export * from './storage/writeObject'
+
 export * from './utils/auth'
 export * from './utils/calculateBasicAuthHeader'
 export * from './utils/calculateBasicAuthUsernamePasswordPair'
@@ -37,10 +40,8 @@ export * from './utils/padHex'
 export * from './utils/path'
 export * from './utils/pkg'
 export * from './utils/plugins'
-export * from './utils/readObject'
 export * from './utils/resolveTree'
 export * from './utils/shasum'
 export * from './utils/sleep'
 export * from './utils/symbols'
 export * from './utils/t'
-export * from './utils/writeObject'
