+++ src/storage/writeObject.js
@@ -1,6 +1,6 @@
 import pako from 'pako'
 
-import { write as writeLoose } from '../managers/GitObjectStoreLoose.js'
+import { writeObjectLoose } from '../storage/writeObjectLoose.js'
 import { FileSystem } from '../models/FileSystem.js'
 import { GitObject } from '../models/GitObject.js'
 import { shasum } from '../utils/shasum.js'
@@ -14,6 +14,6 @@ export async function writeObject ({ fs: _fs, gitdir, type, object, format = 'co
     oid = shasum(object)
     object = Buffer.from(pako.deflate(object))
   }
-  await writeLoose({ fs, gitdir, object, format: 'deflated', oid })
+  await writeObjectLoose({ fs, gitdir, object, format: 'deflated', oid })
   return oid
 }
